+++ packages/apps/tools/src/components/Global/AddNetworkModal/AddNetworkModal.tsx
@@ -11,7 +11,7 @@ import {
 } from '@kadena/react-ui';
 import useTranslation from 'next-translate/useTranslation';
 import type { FC } from 'react';
-import React from 'react';
+import React, {useEffect, useState} from 'react';
 import { useForm } from 'react-hook-form';
 import * as z from 'zod';
 import { formButtonStyle, modalOptionsContentStyle } from './styles.css';
@@ -30,15 +30,15 @@ export const AddNetworkModal: FC<IAddNetworkModalProps> = (props) => {
   const { setSelectedNetwork, setNetworksData, networksData } =
     useWalletConnectClient();
 
-  const {
-    register,
-    handleSubmit: validateThenSubmit,
-    formState: { errors },
-    setError,
-    clearErrors,
-  } = useForm<FormData>({
-    resolver: zodResolver(schema),
-  });
+  const [label, setLabel] = useState('');
+  const [networkId, setNetworkId] = useState('');
+  const [api, setApi] = useState('');
+
+  const [error, setError] = useState('');
+
+  useEffect(() => {
+    setError('');
+  }, [networkId]);
 
   const handleSubmit = (data: FormData, callback: () => void) => {
     const networks = [...networksData];
@@ -47,10 +47,9 @@ export const AddNetworkModal: FC<IAddNetworkModalProps> = (props) => {
     const isDuplicate = networks.find(
       (item) => item.networkId === networkId && item.label === label,
     );
+
     if (isDuplicate) {
-      setError('networkId', {
-        message: 'Error: Duplicate NetworkId',
-      });
+      setError('Error: Duplicate NetworkId');
       return;
     }
 
@@ -65,6 +64,14 @@ export const AddNetworkModal: FC<IAddNetworkModalProps> = (props) => {
     setSelectedNetwork(networkId);
     callback();
   };
+
+  const {
+    register,
+    handleSubmit: validateThenSubmit,
+    formState: { errors },
+  } = useForm<FormData>({
+    resolver: zodResolver(schema),
+  });
 
   return (
     <Dialog {...props}>
@@ -82,35 +89,31 @@ export const AddNetworkModal: FC<IAddNetworkModalProps> = (props) => {
                   <Stack direction="column" gap="$sm">
                     <TextField
                       label={t('Network label')}
-                      inputProps={{
-                        id: 'label',
-                        ...register('label'),
-                        placeholder: 'devnet',
-                      }}
+                      id="label"
+                      {...register('label')}
+                      onChange={(e) => setLabel(e.target.value)}
+                      value={label}
+                      placeholder="devnet"
                       status={errors?.label ? 'negative' : undefined}
                       helperText={errors?.label?.message ?? ''}
                     />
                     <TextField
                       label={t('Network ID')}
-                      inputProps={{
-                        id: 'networkId',
-                        ...register('networkId', {
-                          onChange: () => {
-                            clearErrors('networkId');
-                          },
-                        }),
-                        placeholder: 'fast-development',
-                      }}
+                      id="networkId"
+                      {...register('networkId')}
+                      onChange={(e) => setNetworkId(e.target.value)}
+                      value={networkId}
+                      placeholder="fast-development"
                       status={errors?.networkId ? 'negative' : undefined}
                       helperText={errors?.networkId?.message ?? ''}
                     />
                     <TextField
                       label={t('Network api')}
-                      inputProps={{
-                        id: 'api',
-                        ...register('api'),
-                        placeholder: 'http://localhost:8080',
-                      }}
+                      id="api"
+                      {...register('api')}
+                      onChange={(e) => setApi(e.target.value)}
+                      value={api}
+                      placeholder="localhost:8080"
                       status={errors?.api ? 'negative' : undefined}
                       helperText={errors?.api?.message ?? ''}
                     />