+++ packages/apps/tools/src/components/Global/RequestKeyField/index.tsx
@@ -1,9 +1,8 @@
-import { requestKeyWrapperStyle } from '@/components/Global/RequestKeyField/styles.css';
 import type { ITextFieldProps } from '@kadena/react-ui';
-import { FormFieldWrapper, Input, SystemIcon } from '@kadena/react-ui';
+import { SystemIcon, TextField } from '@kadena/react-ui';
 import useTranslation from 'next-translate/useTranslation';
 import type { FC } from 'react';
-import React from 'react';
+import React, { forwardRef } from 'react';
 import type { FieldError } from 'react-hook-form';
 import * as z from 'zod';
 
@@ -25,41 +24,30 @@ export const REQUEST_KEY_VALIDATION = z
     },
   );
 
-interface IRequestKeyFieldProps
-  extends Partial<Omit<ITextFieldProps, 'inputProps'>> {
-  inputProps: Partial<ITextFieldProps['inputProps']>;
+interface IRequestKeyFieldProps extends Partial<ITextFieldProps> {
   error?: FieldError;
 }
 
-const RequestKeyField: FC<IRequestKeyFieldProps> = ({
-  error,
-  inputProps,
-  status,
-  helperText,
-  ...rest
-}) => {
+const RequestKeyField: FC<IRequestKeyFieldProps> = forwardRef<
+  HTMLInputElement,
+  IRequestKeyFieldProps
+>(function RequestKeyField({ error, status, helperText, ...rest }, ref) {
   const { t } = useTranslation('common');
 
   const helper = helperText || error?.message;
 
   return (
-    <div className={requestKeyWrapperStyle}>
-      <FormFieldWrapper
-        label={t('Request Key')}
-        status={error ? 'negative' : status}
-        helperText={helper}
-        htmlFor={'request-key-input'}
-        {...rest}
-      >
-        <Input
-          {...inputProps}
-          placeholder={t('Enter Request Key')}
-          id={'request-key-input'}
-          startIcon={<SystemIcon.KeyIconFilled />}
-        />
-      </FormFieldWrapper>
-    </div>
+    <TextField
+      ref={ref}
+      label={t('Request Key')}
+      status={error ? 'negative' : status}
+      helperText={helper}
+      id="request-key-input"
+      placeholder={t('Enter Request Key')}
+      startIcon={<SystemIcon.KeyIconFilled />}
+      {...rest}
+    />
   );
-};
+});
 
 export default RequestKeyField;