+++ packages/apps/tools/src/pages/faucet/new/index.tsx
@@ -12,11 +12,9 @@ import {
 } from '@kadena/react-ui';
 
 import {
-  buttonContainerClass,
   hoverTagContainerStyle,
   iconButtonWrapper,
   inputWrapperStyle,
-  notificationContainerStyle,
   notificationContentStyle,
   notificationLinkStyle,
   pubKeyInputWrapperStyle,
@@ -25,9 +23,11 @@ import {
 
 import {
   accountNameContainerClass,
+  buttonContainerClass,
   chainSelectContainerClass,
   containerClass,
   inputContainerClass,
+  notificationContainerStyle,
 } from '../styles.css';
 
 import type { FormStatus } from '@/components/Global';
@@ -302,13 +302,11 @@ const NewAccountFaucetPage: FC = () => {
               <div className={inputWrapperStyle}>
                 <PublicKeyField
                   helperText={errors?.pubKey?.message}
-                  inputProps={{
-                    ...register('pubKey', {
-                      onChange: () => {
-                        clearErrors('pubKey');
-                      },
-                    }),
-                  }}
+                  {...register('pubKey', {
+                    onChange: () => {
+                      clearErrors('pubKey');
+                    },
+                  })}
                   error={errors.pubKey}
                 />
               </div>